feat(refactor): Refactored the code of flowise
